Changed plugin name
Since this was copied from tapCsv, plugin was originally the tapCsv plugin, now it is the tapSpreadsheet plugin to show that it is the main purpose of this particular repo and the tapCsv plugin was discarded.
