+++ src/plugin.ts
@@ -1,138 +1,60 @@
-const through2 = require('through2')
-import Vinyl = require('vinyl')
-import PluginError = require('plugin-error');
-const pkginfo = require('pkginfo')(module); // project package.json info into module.exports
-const PLUGIN_NAME = module.exports.name;
+const through2 = require('through2');
+import PluginError= require('plugin-error')
+import XLSX = require('xlsx')
+const PLUGIN_NAME= 'gulp-etl-tap-spreadsheet'
 import * as loglevel from 'loglevel'
-const log = loglevel.getLogger(PLUGIN_NAME) // get a logger instance based on the project name
-log.setLevel((process.env.DEBUG_LEVEL || 'warn') as log.LogLevelDesc)
+const log = loglevel.getLogger(PLUGIN_NAME)
 
-const parse = require('csv-parse')
-//in theory a better line for this would actually be
-//import {Parser} from 'csv-parse' // WHY YOU NO WORK
-
-/** wrap incoming recordObject in a Singer RECORD Message object*/
 function createRecord(recordObject:Object, streamName: string) : any {
-  return {type:"RECORD", stream:streamName, record:recordObject}
+    return {type:"RECORD", stream:streamName, record:recordObject}
 }
 
-/* This is a gulp-etl plugin. It is compliant with best practices for Gulp plugins (see
-https://github.com/gulpjs/gulp/blob/master/docs/writing-a-plugin/guidelines.md#what-does-a-good-plugin-look-like ),
-and like all gulp-etl plugins it accepts a configObj as its first parameter */
-export function tapCsv(configObj: any) {
-  if (!configObj) configObj = {}
-  if (!configObj.columns) configObj.columns = true // we don't allow false for columns; it results in arrays instead of objects for each record
+//still gotta get html and dbf
 
-  // creating a stream through which each file will pass - a new instance will be created and invoked for each file 
-  // see https://stackoverflow.com/a/52432089/5578474 for a note on the "this" param
-  const strm = through2.obj(function (this: any, file: Vinyl, encoding: string, cb: Function) {
-    const self = this
-    let returnErr: any = null
-    const parser = parse(configObj)
+function convertToJson(file:any){
+    let workbook = XLSX.read(file.contents, {type:"buffer"}) 
+    return XLSX.utils.sheet_to_json(workbook.Sheets[workbook.SheetNames[0]])
+}
 
-    // post-process line object
-    const handleLine = (lineObj: any, _streamName : string): object | null => {
-      if (parser.options.raw || parser.options.info) {
-        let newObj = createRecord(lineObj.record, _streamName)
-        if (lineObj.raw) newObj.raw = lineObj.raw
-        if (lineObj.info) newObj.info = lineObj.info
-        lineObj = newObj
-      }
-      else {
-        lineObj = createRecord(lineObj, _streamName)
-      }
-      return lineObj
-    }
+export function tapSpreadSheet(configObj:any){
+    if (!configObj) configObj = {}
+    if (!configObj.columns) configObj.columns = true
 
-    function newTransformer(streamName : string) {
-
-      let transformer = through2.obj(); // new transform stream, in object mode
-  
-      // transformer is designed to follow csv-parse, which emits objects, so dataObj is an Object. We will finish by converting dataObj to a text line
-      transformer._transform = function (dataObj: Object, encoding: string, callback: Function) {
+    const strm = through2.obj( function (file: any, enc:any, callback:any ){
         let returnErr: any = null
-        try {
-          let handledObj = handleLine(dataObj, streamName)
-          if (handledObj) {
-            let handledLine = JSON.stringify(handledObj)
-            log.debug(handledLine)
-            this.push(handledLine + '\n');
-          }
-        } catch (err) {
-          returnErr = new PluginError(PLUGIN_NAME, err);
+        if ( file.isNull() ){
+            //return empty file
+            return callback(returnErr, file)
         }
-  
-        callback(returnErr)
-      }
-  
-      return transformer
-    }
-
-    // set the stream name to the file name (without extension)
-    let streamName : string = file.stem
-
-    if (file.isNull()) {
-      // return empty file
-      return cb(returnErr, file)
-    }
-    else if (file.isBuffer()) {
-
-
-      parse((file.contents as Buffer), configObj, function (err: any, linesArray: []) {
-        // this callback function runs when the parser finishes its work, returning an array parsed lines 
-        let tempLine: any;
-        let resultArray = [];
-        // we'll call handleLine on each line
-        for (let dataIdx in linesArray) {
-          try {
-            let lineObj = linesArray[dataIdx]
-            tempLine = handleLine(lineObj, streamName)
-            if (tempLine) {
-              let tempStr = JSON.stringify(tempLine)
-              log.debug(tempStr)
-              resultArray.push(tempStr)
+        else if ( file.isStream() ){
+            throw new PluginError(PLUGIN_NAME, 'Trying to stream from unstreamable file')
+        }
+        else if ( file.isBuffer() ){
+            let sheetArr = convertToJson(file)
+            
+            let tempLine: any;
+            let resultArray= [];
+            for (let dataIdx in sheetArr) {
+                try{
+                    let lineObj:any = sheetArr[dataIdx]
+                    tempLine = createRecord(lineObj, 'tap-spreadsheet')
+                    if (tempLine) {
+                        let tempStr = JSON.stringify(tempLine)
+                        log.debug(tempStr)
+                        resultArray.push(tempStr)
+                    }
+                }
+                catch (err){
+                    returnErr = new PluginError(PLUGIN_NAME, err)
+                }
             }
-          }
-          catch (err) {
-            returnErr = new PluginError(PLUGIN_NAME, err)
-          }
+            let data: string = resultArray.join('\n')
+            file.contents = Buffer.from(data);
+            // file.rename({extname:".ndjson"}) 
+            log.debug('calling callback');
+            callback(returnErr, file)
         }
-        let data: string = resultArray.join('\n')
-        file.contents = Buffer.from(data)
+    })
 
-        // we are done with file processing. Pass the processed file along
-        log.debug('calling callback')    
-        cb(returnErr, file);    
-      })
-
-    }
-    else if (file.isStream()) {
-      file.contents = file.contents
-        .pipe(parser)
-        .on('end', function () {
-
-          // DON'T CALL THIS HERE. It MAY work, if the job is small enough. But it needs to be called after the stream is SET UP, not when the streaming is DONE.
-          // Calling the callback here instead of below may result in data hanging in the stream--not sure of the technical term, but dest() creates no file, or the file is blank
-          // cb(returnErr, file);
-          // log.debug('calling callback')    
-
-          log.debug('csv parser is done')
-        })
-        // .on('data', function (data:any, err: any) {
-        //   log.debug(data)
-        // })
-        .on('error', function (err: any) {
-          log.error(err)
-          self.emit('error', new PluginError(PLUGIN_NAME, err));
-        })
-        .pipe(newTransformer(streamName))
-
-        // after our stream is set up (not necesarily finished) we call the callback
-        log.debug('calling callback')    
-        cb(returnErr, file);
-    }
-
-  })
-
-  return strm
+    return strm;
 }
